refactor(ingestion): Updated Redshift lineage_v1 refs and removed v1 implementation (#14580)
